requirements.txt
asgiref==3.7.2
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
et-xmlfile==1.1.0
idna==3.6
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
